+++ civis/io/_files.py
@@ -122,7 +122,7 @@ def _multipart_upload(buf, name, file_size, client, **kwargs):
     num_parts = int(math.ceil((file_size) / float(part_size)))
 
     log.debug('Uploading file with %s bytes using %s file parts with a part '
-              'size of %s bytes', file_size, part_size, num_parts)
+              'size of %s bytes', file_size, num_parts, part_size)
     file_response = client.files.post_multipart(name=name, num_parts=num_parts,
                                                 **kwargs)