+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Updated CivisML 2.0 notebook (#214)
+- Reworded output of `civis notebooks new` CLI command (#215)
 
 ## 1.8.0 - 2018-01-23
 ### Added