+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 ### Changed
-
+- Updated CivisML 2.0 notebook (#214)
 - Reworded output of `civis notebooks new` CLI command (#215)
 
 ## 1.8.0 - 2018-01-23